Fix a warning C4018 (#322)
* Fix a warning C4018

https://github.com/fuzzball-muck/fuzzball/blob/master/src/look.c#L1084
https://ci.appveyor.com/project/fuzzball-muck/fuzzball/builds/22582572#L265
This will likely introduce new warnings, which I'll have to check via CI since I don't have access to Windows.  I also fixed minor formatting/alignment. :P

* Fix formatting

Uses tabs instead of spaces for the indentation, which are far easier to indent with.
The file also uses tabs for aligning the comments though...  That I consider more dubious, BUT, it was already like that and that's not my change.
I also switched the comment after the endif to only be separated by a single space, as per the only styleguide I read which had anything to say about it.
https://www.freebsd.org/cgi/man.cgi?query=style&apropos=0&sektion=9&manpath=FreeBSD+12.0-RELEASE+and+Ports&arch=default&format=html

+++ include/dbsearch.h
@@ -2,31 +2,31 @@
 #define _DBSEARCH_H
 
 struct flgchkdat {
-    int fortype;		/* check FOR a type? */
-    int istype;			/* If check FOR a type, which one? */
-    int isnotroom;		/* not a room. */
-    int isnotexit;		/* not an exit. */
-    int isnotthing;		/* not type thing */
-    int isnotplayer;		/* not a player */
-    int isnotprog;		/* not a program */
-    int forlevel;		/* check for a mucker level? */
-    int islevel;		/* if check FOR a mucker level, which level? */
-    int isnotzero;		/* not ML0 */
-    int isnotone;		/* not ML1 */
-    int isnottwo;		/* not ML2 */
-    int isnotthree;		/* not ML3 */
-    int setflags;		/* flags that are set to check for */
-    int clearflags;		/* flags to check are cleared. */
-    int forlink;		/* check linking? */
-    int islinked;		/* if yes, check if not unlinked */
-    int forold;			/* check for old object? */
-    int isold;			/* if yes, check if old */
-    int loadedsize;		/* check for propval-loaded size? */
-    int issize;			/* list objs larger than size? */
-    int size;			/* what size to check against. No check if 0 */
+	int fortype;		/* check FOR a type? */
+	int istype;		/* If check FOR a type, which one? */
+	int isnotroom;		/* not a room. */
+	int isnotexit;		/* not an exit. */
+	int isnotthing;		/* not type thing */
+	int isnotplayer;	/* not a player */
+	int isnotprog;		/* not a program */
+	int forlevel;		/* check for a mucker level? */
+	int islevel;		/* if check FOR a mucker level, which level? */
+	int isnotzero;		/* not ML0 */
+	int isnotone;		/* not ML1 */
+	int isnottwo;		/* not ML2 */
+	int isnotthree;		/* not ML3 */
+	int setflags;		/* flags that are set to check for */
+	int clearflags;		/* flags to check are cleared. */
+	int forlink;		/* check linking? */
+	int islinked;		/* if yes, check if not unlinked */
+	int forold;		/* check for old object? */
+	int isold;		/* if yes, check if old */
+	int loadedsize;		/* check for propval-loaded size? */
+	int issize;		/* list objs larger than size? */
+	size_t size;		/* what size to check against. No check if 0 */
 };
 
 int checkflags(dbref what, struct flgchkdat check);
 int init_checkflags(dbref player, const char *flags, struct flgchkdat *check);
 
-#endif				/* _DBSEARCH_H */
+#endif /* _DBSEARCH_H */
